Move determineExecutionPlan out of TaskExecutionPlan
